Refactor LLM API structure: move LLMApi and related classes to core module, update imports, and adjust initialization logic
